refactor: refactor api for manage lessons
